requirements.txt
python-pptx
pytesseract
Pillow
regex
requests
numpy
